Drop tmate for ngrok and Tor; add collaboration
Tmate did not adequately support the necessary features of tmux
required for session groups that would allow proper collaboration.
Instead we use ttyd and/or an SSH server with an actual tmux
process configured the way we need.

For more information on the shortcomings of tmate in this
particular case, see (among others):
https://github.com/tmate-io/tmate/issues/49